# Conflicts:
#	stan/math/rev/fun/columns_dot_self.hpp
#	stan/math/rev/fun/dot_product.hpp
#	test/expressions/stan_math_sigs_exceptions.expected

+++ test/unit/math/mix/fun/cholesky_decompose_test.cpp
@@ -129,7 +129,8 @@ TEST(MathMixMatFun, choleskyDecomposeGP) {
       for (int i = 0; i < n; ++i) {
         Sigma(i, i) = 1;
         for (int j = 0; j < i; ++j) {
-          Sigma(i, j) = std::pow(rho, fabs(i - j));
+          Sigma(i, j) = std::pow(
+              rho, fabs(static_cast<double>(i) - static_cast<double>(j)));
           Sigma(j, i) = Sigma(i, j);
         }
       }
